fixed MachineDaisyFile, updated EZ_DSP.h,  updated kickDrum class, changed JOEoscillator.cpp to .h
changed sample rate to 32khz to compensate for slow snare class. Moved button ladder stuff to AudioCallback and added a simple debounce boolean. Need to build our own snare that processes faster so that we can increase the sample rate if necessary. kickDrum needs to be fixed, it was breaking. JOEoscillator might have to be deprecated and replaced with oscillator class in kickDrum